apps/phoenix-rise/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
astroid==3.3.11
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.1
backoff==2.2.1
bandit==1.8.6
bcrypt==4.3.0
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
boltons==21.0.0
boolean.py==5.0
bracex==2.6
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.1.8
click-option-group==0.5.7
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
crewai==0.150.0
cryptography==45.0.5
cyclonedx-python-lib==9.1.0
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
dparse==0.6.4
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
face==24.0.0
filelock==3.16.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
glom==22.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==7.1.0
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.10.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jellyfish==1.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.4.8
language_data==1.3.0
license-expression==30.4.4
litellm==1.74.3
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
murmurhash==1.0.13
mypy==1.17.0
mypy_extensions==1.1.0
narwhals==1.47.1
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.97.1
openpyxl==3.1.5
openrouter==1.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.11.1
overrides==7.7.0
packageurl-python==0.17.3
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pbr==6.1.1
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.2
pexpect==4.9.0
# Editable install with no version control (phoenix-shared-auth==1.0.0)
-e /Users/mattvaness/Desktop/IA/phoenix/phoenix-eco/phoenix_shared_auth
# Editable install with no version control (phoenix_shared_models==0.1.0)
-e /Users/mattvaness/Desktop/IA/phoenix/phoenix-eco/phoenix_shared_models
phonenumbers==8.13.55
pika==1.3.2
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
pipreqs==0.4.13
platformdirs==4.3.8
playwright==1.54.0
plotly==6.2.0
pluggy==1.6.0
portalocker==2.7.0
postgrest==1.1.1
posthog==5.4.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-serializable==2.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic-settings==2.10.1
pydantic_core==2.23.4
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.403
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
realtime==2.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.5.3
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.4
safety==3.6.0
safety-schemas==0.0.14
segtok==1.5.11
semgrep==1.131.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.42
srsly==2.5.1
stack-data==0.6.3
stevedore==5.4.1
storage3==0.12.0
streamlit==1.47.0
StrEnum==0.4.15
stripe==12.4.0
structlog==25.4.0
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.0.2
tomli_w==1.2.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.3
uvicorn==0.35.0
uvloop==0.21.0
vulture==2.14
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcmatch==8.5.2
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yake==0.6.0
yarg==0.1.10
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
